Foto Owl/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
alembic
pydantic
python-jose
passlib[bcrypt]==1.7.4
python-multipart